Weighted sectors
Every sector is now defined not only by its action, but also a weight.
A sector with weight N*k is N times more likely to be landed on than a
sector with weight k.

Also use a single <defn><path> element with a sector slice, then
reference it with <use> and an appropriate transform matrix for all
sectors.